API, Redis, Mosquitto, heartbeat
API done
Redis done
mosquitto publisher method, draft of subscriber thread
heartbeat master method done, draft of slaves process

+++ core/API/nodes.py
@@ -1,28 +1,41 @@
 #!thesis/api
 
-from flask import abort, make_response, jsonify
-from core.common import Node
-from core.database import utils
-import uuid
+from flask import make_response, jsonify
+from core.databaseRedis import nodeUtils as utils
+from validator import validateNodeRequest as validate
+from core.utils import uniqueName
 
 def newNode(request):
-    req = request.json
-
-    if not request.json:
-        abort(400)
-    if not 'ip' in req or type(req['ip']) != unicode:
-        return make_response(jsonify({'error': 'no ip'}), 400)
-    if not "architecture" in req or type(req['architecture']) != unicode:
-        return make_response(jsonify({'error': 'no arch'}), 400)
-
-    node = Node(uuid.uuid4(), req['ip'], req['architecture'])
     
-    return make_response(jsonify(node.__dict__), 200)
+    valid, resp = validate(request)
+    if not valid:
+        return make_response(jsonify(resp), 400)
+
+    id = uniqueName()
+    resp['id'] = id
+    resp['mqtt_Topic'] = "node/" + id
+    
+    if resp.pop('setup'):
+        user = resp.pop("ssh_user")
+        password = resp.pop ("ssh_password")
+
+        #TODO RUN SETUP
+
+    utils.insertNode(resp)
+
+    return make_response(id, 200)
+
 
 def deleteNode(request, token):
     
-    res = utils.deleteNode(token)
-    
-    if not res:
-        abort(404)
+    if not utils.getNode(token):
+        return make_response(jsonify({'error': "No node with id " + token}), 406)
 
+    utils.deleteNode(token)
+    
+    return make_response("OK", 200)
+
+
+def getNodes(request):
+    nodes = utils.getNodes()
+    return make_response(jsonify({"nodes": nodes}), 200)
